Adds a way to debug the sneakers integration

+++ spec/support/file_helper.rb
@@ -23,6 +23,6 @@ module FileHelper
   end
 
   def delete_file(file_path)
-    File.delete(file_path)
+    File.delete(file_path) if File.exist?(file_path)
   end
 end
